import: use ops to create objects
Create the armature object using bpy.ops. This has the advantage
of placing it at the 3D Cursor, which is normal behaviour for
an op, and also of having Blender do linking for us (linking
changed in Blender 2.8).

Despite the comment near make_model(), models without an armature
did not have a root Empty created for them. This was also added
and created with an op the same way.