import: add actions to NLA track
All the animation actions are pushed to one NLA track on the
armature object, one after another. This means you can now see
all the animations in an imported model by just hitting play.

Fake users are no longer required for the actions.

The repeat count is set to 2 for looping animations. This is used
by the EXM exporter to know when to set the looping flag.